@@ -1,26 +1,26 @@
+using BlazorApp.Models;
+using Microsoft.AspNetCore.Components;
 using System.Net.Http.Json;
 
-using BlazorApp.Models;
+namespace BlazorApp.Shared;
 
-using Microsoft.AspNetCore.Components;
-
-namespace BlazorApp.Shared
+public partial class Footer
 {
-    public partial class Footer
-    {
-        [Parameter, EditorRequired]
-        public HttpClient Http { get; set; }
+    [Parameter, EditorRequired]
+    public required HttpClient Http { get; set; }
 
-        [Parameter]
-        public string PrimaryColor { get; set; } = string.Empty;
-        [Parameter]
-        public string SecondaryColor { get; set; } = string.Empty;
-        private SiteProperties? property;
-        private SocialIcons? icons;
-        protected override async Task OnInitializedAsync()
-        {
-            property = await Http.GetFromJsonAsync<SiteProperties>("sample-data/siteproperties.json");
-            icons = await Http.GetFromJsonAsync<SocialIcons>("sample-data/socialicons.json");
-        }
+    [Parameter]
+    public string PrimaryColor { get; set; } = string.Empty;
+
+    [Parameter]
+    public string SecondaryColor { get; set; } = string.Empty;
+
+    private SiteProperties? property;
+    private SocialIcons? icons;
+
+    protected override async Task OnInitializedAsync()
+    {
+        property = await Http.GetFromJsonAsync<SiteProperties>("sample-data/siteproperties.json");
+        icons = await Http.GetFromJsonAsync<SocialIcons>("sample-data/socialicons.json");
     }
 }
